refactor: 베팅 결과 메서드 이동
