+++ src/main/kotlin/blackjack/model/participant/Participants.kt
@@ -1,8 +1,9 @@
 package blackjack.model.participant
 
+import blackjack.model.betting.BettingResult
+import blackjack.model.betting.BettingTable
 import blackjack.model.card.Card
 import blackjack.model.card.CardCount
-import blackjack.model.winning.PlayersResult
 
 class Participants(
     val dealer: Dealer,
@@ -14,7 +15,41 @@ class Participants(
         }
     }
 
-    val playersResult: PlayersResult get() = players.winningResult(dealer)
+    fun betMoney(setBettingMoney: (Name) -> Money): BettingTable {
+        val bettingTable = BettingTable()
+        players.value.forEach { player ->
+            val bettingMoney = setBettingMoney(player.name)
+            require(bettingMoney > Money.ZERO) {
+                ("[ERROR] 베팅 금액은 0원보다 높아야 합니다.")
+            }
+
+            player.payMoney(bettingMoney)
+
+            bettingTable.add(player.name, bettingMoney)
+        }
+        return bettingTable
+    }
+
+    fun profitResult(bettingTable: BettingTable): BettingResult {
+        val resultBettingTable = BettingTable()
+        updateTable(bettingTable, resultBettingTable)
+        return resultBettingTable.table
+    }
+
+    private fun updateTable(
+        previousBettingTable: BettingTable,
+        resultBettingTable: BettingTable,
+    ) {
+        players.winningResult(dealer).value.forEach { (name, result) ->
+            val profit = previousBettingTable.get(name) * result.profitRate
+
+            resultBettingTable.add(dealer.name, -profit)
+            resultBettingTable.add(name, profit)
+
+            dealer.receiveMoney(-profit)
+            players.receiveMoney(name, profit)
+        }
+    }
 
     companion object {
         fun create(
